Refactor: services content
