+++ src/hooks/useNavigationLinks.tsx
@@ -1,28 +1,31 @@
-import { useEffect, useState } from 'react'
+import { useEffect, useState } from "react";
 
 export function useNavigationLinks() {
-  const [isDashboard, setIsDashboard] = useState(false)
+  const [isDashboard, setIsDashboard] = useState(false);
 
   useEffect(() => {
-    const hostname = window.location.hostname
-    setIsDashboard(hostname.startsWith('dashboard.') || hostname === 'dashboard.localhost')
-  }, [])
+    const hostname = window.location.hostname;
+    setIsDashboard(
+      hostname.startsWith("dashboard.") || hostname === "dashboard.localhost",
+    );
+  }, []);
 
   const publicLinks: NavigationLink[] = [
-    { to: '/simulados', label: 'Simulados' },
-    { to: '/questoes', label: 'Questões' },
-    { to: '/conteudos', label: 'Conteúdos' },
-    { to: '/provas', label: 'Provas' },
-    { to: '/turmas', label: 'Turmas' },
-  ]
+    { to: "/simulados", label: "Simulados" },
+    { to: "/questoes", label: "Questões" },
+    { to: "/conteudos", label: "Conteúdos" },
+    { to: "/provas", label: "Provas" },
+    { to: "/turmas", label: "Turmas" },
+  ];
 
   const dashboardLinks: NavigationLink[] = [
-    { to: '/questoes', label: 'Questões' },
-    { to: '/content/contents', label: 'Conteúdos' },
-    { to: '/usuarios', label: 'Usuários' },
-  ]
+    { to: "/question", label: "Questões" },
+    { to: "/content/contents", label: "Conteúdos" },
+    { to: "/usuarios", label: "Usuários" },
+  ];
 
-  const links: NavigationLink[] = isDashboard ? dashboardLinks : publicLinks
+  const links: NavigationLink[] = isDashboard ? dashboardLinks : publicLinks;
 
-  return { links, isDashboard }
+  return { links, isDashboard };
 }
+
